+++ sklearn/exact_tree/_tree_builder.pxd
@@ -4,8 +4,8 @@ from ._tree cimport TreeBuilder
 from ._splitter cimport Splitter
 
 
-cdef void weighted_sum_y(double[::1] y, double[::1] sample_weight,
-                         double* p_sum_y, double* p_sum_sq_y)
+cdef void weighted_sum_y(float[::1] y, float[::1] sample_weight,
+                         float* p_sum_y, float* p_sum_sq_y)
 
 
 cdef class ExactTreeBuilder(TreeBuilder):
@@ -13,14 +13,14 @@ cdef class ExactTreeBuilder(TreeBuilder):
     cdef:
         int min_samples_split
         int min_samples_leaf
-        double min_weight_leaf
-        double min_impurity_split
+        float min_weight_leaf
+        float min_impurity_split
         int max_depth
         int max_features
 
     cpdef build(self, Tree tree, float[:, ::1] X, int[::1, :] X_idx_sorted,
-                double[::1] y, double[::1] sample_weight,
-                double sum_total_weighted_samples)
+                float[::1] y, float[::1] sample_weight,
+                float sum_total_weighted_samples)
 
     cdef inline bint _check_minimum_stats(self, Splitter* splitter):
         cdef:
